fix: green blockers not appearing because of async fetch blocking the flow
