+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/RendereableAssetLoadHelper/RendereableAssetLoadHelper.asmdef
@@ -7,11 +7,12 @@
         "GUID:ee6523961ef177343ad763fcd55c7c46",
         "GUID:fbcc413e192ef9048811d47ab0aca0c0",
         "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
-        "GUID:18ee28afad71091499d9ebad494c995e",
         "GUID:2995626b54c60644988f134a69a77450",
         "GUID:28f74c468a54948bfa9f625c5d428f56",
         "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
-        "GUID:2fe81fa47ac9ca345860b544ce917241"
+        "GUID:2fe81fa47ac9ca345860b544ce917241",
+        "GUID:26861d1714ac819459758e0b653efd06",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
